% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_no_diacritic.R
\name{add_no_diacritic}
\alias{add_no_diacritic}
\title{Get Subdivision Names without Diacritics}
\usage{
add_no_diacritic(adf, merge = TRUE)
}
\arguments{
\item{adf}{A data frame with a \code{subdivision_name} column}

\item{merge}{Merge with the previous data? Defaults to TRUE}
}
\value{
a tibble
}
\description{
Get Subdivision Names without Diacritics
}
\examples{
\dontrun{
iso_31662_subdivisions_all <- add_no_diacritic(iso_31662_subdivisions)

}
}
