% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subdivisions.R
\name{clean_subdivisions}
\alias{clean_subdivisions}
\title{Create Clean Subdivision Names without Additional Regional Specifications}
\usage{
clean_subdivisions(adf, merge = TRUE)
}
\arguments{
\item{adf}{A data frame with a \code{subdivision_name} column}

\item{merge}{Merge with the previous data? Defaults to TRUE}
}
\value{
a tibble
}
\description{
Create Clean Subdivision Names without Additional Regional Specifications
}
\examples{
\dontrun{
iso_31662_subdivisions_clean <- clean_subdivisions(iso_31662_subdivisions)

}
}
