% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_abbreviations.R
\name{expand_abbreviations}
\alias{expand_abbreviations}
\title{Expand common abbreviations in names}
\usage{
expand_abbreviations(adf, merge = TRUE)
}
\arguments{
\item{adf}{A data frame with a \code{subdivision_name} column}

\item{merge}{Merge with the previous data? Defaults to TRUE}
}
\value{
A tibble of information of codes in the same format as \code{\link{iso_31662_subdivisions}}
}
\description{
Expand common abbreviations in names
}
\examples{
\dontrun{
iso_31662_subdivisions_expanded <- expand_abbreviations(iso_31662_subdivisions)

}
}
