% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso_31662_countries}
\alias{iso_31662_countries}
\title{ISO 3166-2 Country Codes from Wikipedia}
\format{
A data frame with 249 rows and 2 variables:
\describe{
\item{country_code}{The ISO 3166-2 Country Code of that nation}
\item{country_name}{The name of the country (english)}
...
}
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
\usage{
iso_31662_countries
}
\description{
A dataset of country names and their accompanying ISO 3166 2 codes scraped from
Wikipedia
}
\keyword{datasets}
