% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso_31662_subdivisions}
\alias{iso_31662_subdivisions}
\title{ISO 3166-2 Subdivision Codes from Wikipedia}
\format{
A data frame with 7784 rows and 4 variables:
\describe{
\item{code}{The ISO 3166-2 Subdivision Code}
\item{country_code}{The ISO 3166-2 Country Code of the country the subdivision is in}
\item{subdivision_name_type}{The type of name (from Wikipedia column). Many
subdivisions have their names in multiple languages}
\item{subdivision_name}{The name of the subdivision in that type}
}
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
\usage{
iso_31662_subdivisions
}
\description{
A dataset many variations of subdivision names and their accompanying ISO 3166 2 codes
}
\keyword{datasets}
