% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso_31662_subdivisions_extended}
\alias{iso_31662_subdivisions_extended}
\title{ISO 3166-2 Subdivision Codes from Wikipedia with Additional Modified Subdivisions}
\format{
A data frame with 10349 rows and 4 variables:
\describe{
\item{code}{The ISO 3166-2 Subdivision Code}
\item{country_code}{The ISO 3166-2 Country Code of the country the subdivision is in}
\item{subdivision_name_type}{The type of name (from Wikipedia column). Many
subdivisions have their names in multiple languages}
\item{subdivision_name}{The name of the subdivision in that type}
}
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
\usage{
iso_31662_subdivisions_extended
}
\description{
Many different subdivision name datasets deviate from what is in the ISO standard.
This dataset seeks to provide a more comprehensive listing of names for easier
acquisition of standardized ISO 3166-2 subdivision codes via a merge or join operation.
Currently includes
\describe{
\item{}{Names with no diacritics}
\item{}{Names cleaned of status signifiers (e.g., Township, Capital, etc.)}
\item{}{Names from iso-3166-2.js}
}
}
\keyword{datasets}
