% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_countries.R
\name{scrape_countries}
\alias{scrape_countries}
\title{Scrape ISO 3166-2 Country Codes from Wikipedia}
\usage{
scrape_countries()
}
\value{
A tibble of country codes and country names
}
\description{
Scrape ISO 3166-2 Country Codes from Wikipedia
}
\examples{
\dontrun{
iso_countries <- scrape_countries()

head(iso_countries)
}
}
\references{
Wikipedia ISO-3166-2 Entry: \url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
