% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_iso_3166_2_js.R
\name{scrape_iso_3166_2_js}
\alias{scrape_iso_3166_2_js}
\title{Get the ISO 3166 2 Subdivision Table from iso-3166-2.js}
\usage{
scrape_iso_3166_2_js()
}
\value{

}
\description{
Get the ISO 3166 2 Subdivision Table from iso-3166-2.js
}
\examples{
\dontrun{
scrape_iso_3166_2_js() \%>\%
    head()
}
}
