% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rnaturalearth.R
\name{scrape_rnaturalearth}
\alias{scrape_rnaturalearth}
\title{Scrape ISO 3166-2 Subdivision Codes from RNaturalearth}
\usage{
scrape_rnaturalearth()
}
\value{
A tibble of information of codes, types of names for a subdivision in
in the same column format as iso_31662_subdivisions
}
\description{
Uses the ropensci rnaturalearth package to get moar ISO 3166-2 codes
}
\examples{
\dontrun{
iso_rinat <- scrape_rnaturalearth()

head(iso_rinat)
}
}
\references{
\url{https://www.naturalearthdata.com/}

\url{https://github.com/ropensci/rnaturalearth}
}
