% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_subdivisions.R
\name{scrape_subdivisions}
\alias{scrape_subdivisions}
\title{Scrape ISO 3166-2 Subdivision Codes from Wikipedia}
\usage{
scrape_subdivisions(quiet = TRUE)
}
\arguments{
\item{quiet}{Should the debug information be quieted? Defaults to TRUE.}
}
\value{
A tibble of information of codes, types of names for a subdivision in
Wikipedia, and it's name.
}
\description{
Scrape ISO 3166-2 Subdivision Codes from Wikipedia
}
\examples{
\dontrun{
iso_sub <- scrape_subdivisions()

head(iso_sub)
}
}
\references{
Wikipedia ISO-3166-2 Entry: \url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
